+++ gems/aws-sdk-sqs/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Raise ArgumentError when QueuePoller's :max_number_of_messages is not a positive integer.
+
 1.52.0 (2022-10-25)
 ------------------